Se agregó un botón genérico que sirve para guardar y que carga
